Add custom exception handling and null checks
- Introduced `NotFoundExceptions` for handling missing leave types.
- Added `BadRequsetExceptions` for bad request scenarios.
- Updated `DeleteLeaveTypeCommandHandler` and `GetLeaveTypeDetailsQueryHandler` to check for null leave types.
- Standardized parameter naming in `GetLeaveTypeDetailsQuery`.

+++ HRLeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HRLeaveManagement.Application.Contracts.Persistence;
+using HRLeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HRLeaveManagement.Application.Features.LeaveType.Commands.CreateLeaveType;
@@ -15,6 +16,10 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveType = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
+        if (leaveType == null)
+            throw new NotFoundExceptions(nameof(LeaveType), request.Id);
+
         await _leaveTypeRepository.DeleteAsync(leaveType);
 
         return Unit.Value;
